move genomic enums
