+++ rdr_service/alembic/versions/50d9eeb498c3_genomic_contamination_records.py
@@ -14,7 +14,8 @@ from rdr_service.participant_enums import PhysicalMeasurementsStatus, Questionna
 from rdr_service.participant_enums import WithdrawalStatus, WithdrawalReason, SuspensionStatus, QuestionnaireDefinitionStatus
 from rdr_service.participant_enums import EnrollmentStatus, Race, SampleStatus, OrganizationType, BiobankOrderStatus
 from rdr_service.participant_enums import OrderShipmentTrackingStatus, OrderShipmentStatus
-from rdr_service.participant_enums import MetricSetType, MetricsKey, GenderIdentity, GenomicJob
+from rdr_service.participant_enums import MetricSetType, MetricsKey, GenderIdentity
+from rdr_service.genomic_enums import GenomicJob
 from rdr_service.model.base import add_table_history_table, drop_table_history_table
 from rdr_service.model.code import CodeType
 from rdr_service.model.site_enums import SiteStatus, EnrollingStatus, DigitalSchedulingStatus, ObsoleteStatus
