+++ rdr_service/alembic/versions/e8df4ef80f31_create_genomic_centers_tables.py
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 from sqlalchemy.dialects import mysql
 
 from rdr_service.model import utils
-from rdr_service.participant_enums import GenomicSubProcessStatus, GenomicSubProcessResult
+from rdr_service.genomic_enums import GenomicSubProcessStatus, GenomicSubProcessResult
 
 # revision identifiers, used by Alembic.
 revision = 'e8df4ef80f31'
@@ -135,7 +135,7 @@ def upgrade_rdr():
     # create jobs
     op.execute(
         """
-        INSERT INTO genomic_job (name, active_flag) 
+        INSERT INTO genomic_job (name, active_flag)
         VALUES ("gc_validation_metrics_ingestion", 1);
         """
     )
