+++ rdr_service/genomic/genomic_set_file_handler.py
@@ -13,7 +13,7 @@ from rdr_service import clock, config
 from rdr_service.api_util import open_cloud_file, list_blobs
 from rdr_service.dao.genomics_dao import GenomicSetDao, GenomicSetMemberDao
 from rdr_service.model.genomics import GenomicSet, GenomicSetMember
-from rdr_service.participant_enums import GenomicSetStatus, GenomicSetMemberStatus
+from rdr_service.genomic_enums import GenomicSetStatus, GenomicSetMemberStatus
 from rdr_service.offline.sql_exporter import SqlExporter
 
 _US_CENTRAL = pytz.timezone("US/Central")
@@ -260,7 +260,7 @@ def _create_and_upload_result_file(genomic_set):
     bucket_name = config.getSetting(config.GENOMIC_SET_BUCKET_NAME)
     exporter = SqlExporter(bucket_name)
     export_sql = """
-    SELECT 
+    SELECT
       :genomic_set_name AS genomic_set_name,
       :genomic_set_criteria AS genomic_set_criteria,
       participant_id AS pid,
